Fix -Wdouble-promotion for rating in PLT_MediaObject::ToDidl
fixes:
../lib/libUPnP/Platinum/Source/Devices/MediaServer/PltMediaItem.cpp: In member function 'virtual NPT_Result PLT_MediaObject::ToDidl(NPT_UInt64, NPT_String&)':
../lib/libUPnP/Platinum/Source/Devices/MediaServer/PltMediaItem.cpp:589:55: warning: implicit conversion from 'NPT_Float' {aka 'float'} to 'double' when passing argument to function [-Wdouble-promotion]
  589 |         didl += NPT_String::Format("%.1f", m_XbmcInfo.rating);
      |                                            ~~~~~~~~~~~^~~~~~

+++ lib/libUPnP/Platinum/Source/Devices/MediaServer/PltMediaItem.cpp
@@ -586,7 +586,7 @@ PLT_MediaObject::ToDidl(NPT_UInt64 mask, NPT_String& didl)
     // xbmc rating
     if (mask & PLT_FILTER_MASK_XBMC_RATING) {
         didl += "<xbmc:rating>";
-        didl += NPT_String::Format("%.1f", m_XbmcInfo.rating);
+        didl += NPT_String::Format("%.1f", (double)m_XbmcInfo.rating);
         didl += "</xbmc:rating>";
     }
